Organized the files into different folders coresponding to what they are
