+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Rosen Raikov on 6.10.24.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
