+++ FitnessApp/Home/Views/HomeView.swift
@@ -7,27 +7,6 @@
 
 import SwiftUI
 
-class HomeViewModel: ObservableObject {
-    
-    @State var calories: Int = 123
-    @State var active: Int = 50
-    @State var stand: Int = 8
-    
-    @Published var mockActivities = [
-        Activity(id: 0, title: "Todays Steps", subtitle: "Goal: 10,000", image: "figure.walk", tintColor: .green, amount: "6,121"),
-        Activity(id: 1, title: "Calories", subtitle: "Goal: 600", image: "flame", tintColor: .red, amount: "250"),
-        Activity(id: 2, title: "Stand", subtitle: "Goal: 8 hours", image: "figure.stand", tintColor: .blue, amount: "5 hours"),
-        Activity(id: 3, title: "Stand", subtitle: "Goal: 8 hours", image: "figure.stand", tintColor: .blue, amount: "5 hours")
-    ]
-    
-    @Published var mockWorkouts = [
-        Workout(id: 0, title: "Running", image: "figure.run", tintColor: .green, duration: "23 mins", date: "Aug 19", calories: "341 kcal"),
-        Workout(id: 1, title: "Boxing", image: "figure.boxing", tintColor: .red, duration: "2 hours", date: "Aug 10", calories: "1000 kcal"),
-        Workout(id: 2, title: "Skiing", image: "figure.skiing.downhill", tintColor: .blue, duration: "4 hours", date: "Aug 10", calories: "550 kcal"),
-        Workout(id: 3, title: "Football", image: "figure.indoor.soccer", tintColor: .orange, duration: "45 mins", date: "Aug 25", calories: "870 kcal")
-    ]
-}
-
 struct HomeView: View {
     
     @StateObject var viewModel = HomeViewModel()
@@ -42,7 +21,6 @@ struct HomeView: View {
                     
                     HStack {
                         
-                        
                         Spacer()
                         
                         VStack {
@@ -52,7 +30,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(Color.red)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.calories)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -63,7 +41,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(Color.green)
                                 
-                                Text("50 mins")
+                                Text("\(viewModel.active)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -74,7 +52,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(Color.blue)
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand)")
                                     .bold()
                             }
                         }
@@ -82,12 +60,12 @@ struct HomeView: View {
                         Spacer()
                         
                         ZStack {
-                            ProgressCricleView(progress: viewModel.$calories, goal: 600, color: .red)
+                            ProgressCricleView(progress: $viewModel.calories, goal: 600, color: .red)
                             
-                            ProgressCricleView(progress: viewModel.$active, goal: 120, color: .green)
+                            ProgressCricleView(progress: $viewModel.active, goal: 120, color: .green)
                                 .padding(.all, 20)
                             
-                            ProgressCricleView(progress: viewModel.$stand, goal: 12, color: .blue)
+                            ProgressCricleView(progress: $viewModel.stand, goal: 12, color: .blue)
                                 .padding(.all, 40)
                             
                         }
